32k Clicking

+++ src/main/java/com/runescape/entity/GameObject.java
@@ -17,9 +17,6 @@ public final class GameObject {
     public int yLocLow;
     public int anInt527;
     public int anInt528;
-    public int uid;
-    /**
-     * mask = (byte)((objectRotation << 6) + objectType);
-     */
-    public byte mask;
+    public long uid;
+    public int mask;
 }
